LayoutTests/inspector/canvas/create-canvas-contexts-expected.txt

Test that CanvasManager tracks creation and destruction of 2D and WebGL canvases.


== Running test suite: Canvas.canvasContexts
-- Running test case: CheckNoCanvases
PASS: CanvasManager should have no canvases.

-- Running test case: CheckCanvasesCleared
Added canvas.
Cleared canvases after reloading page.
PASS: CanvasManager should have no canvases.

-- Running test case: Create2DCanvasContext
Added canvas.
PASS: Canvas context should be 2D.
Removed canvas.
PASS: Removed canvas has expected ID.

-- Running test case: CreateWebGLCanvasContext
Added canvas.
PASS: Canvas context should be WebGL.
Removed canvas.
PASS: Removed canvas has expected ID.

-- Running test case: Create2DCanvasContextOffscreen
Added canvas.
PASS: Canvas context should be 2D.
Removed canvas.
PASS: Removed canvas has expected ID.

-- Running test case: CreateWebGLCanvasContextOffscreen
Added canvas.
PASS: Canvas context should be WebGL.
Removed canvas.
PASS: Removed canvas has expected ID.

-- Running test case: CreateCSSCanvas2D
Create CSS canvas from -webkit-canvas(css-canvas-2d).
Added canvas.
PASS: Canvas context should be 2D.
PASS: Canvas name should equal the identifier passed to -webkit-canvas.

-- Running test case: CreateCSSCanvasWebGL
Create CSS canvas from -webkit-canvas(css-canvas-webgl).
Added canvas.
PASS: Canvas context should be WebGL.
PASS: Canvas name should equal the identifier passed to -webkit-canvas.
